+++ Arduino/HandTieArduinoSystem/StrainGauge.h
@@ -7,7 +7,7 @@ class StrainGauge
 {
 public:
    StrainGauge();
-   StrainGauge(uint8_t, uint8_t, uint8_t, uint8_t);
+   StrainGauge(uint8_t, uint8_t, uint16_t, uint16_t);
    ~StrainGauge();
 
    void setAmpPotPos(uint8_t);
@@ -16,18 +16,18 @@ public:
    void setBridgePotPos(uint8_t);
    uint8_t getBridgePotPos();
 
-   void setTargetValNoAmp(uint8_t);
-   uint8_t getTargetValNoAmp();
+   void setTargetValNoAmp(uint16_t);
+   uint16_t getTargetValNoAmp();
 
-   void setTargetValWithAmp(uint8_t);
-   uint8_t getTargetValWithAmp();
+   void setTargetValWithAmp(uint16_t);
+   uint16_t getTargetValWithAmp();
 
 private:
    uint8_t ampPotPos;
    uint8_t bridgePotPos;
 
-   uint8_t targetValNoAmp;
-   uint8_t targetValWithAmp;
+   uint16_t targetValNoAmp;
+   uint16_t targetValWithAmp;
 };
 
 #endif   //StrainGauge_h
